Refactored code, put files into website, refined login/signup logic
