--- package.json
@@ -1,25 +1,0 @@
-{
-  "name": "david_vacationworld",
-  "version": "1.0.0",
-  "description": "TripPin is a web application that will allow you to map out places in the world that you have been to,\r ability to plan for next vacation ideas, and itineraries, provide ratings and reviews to\r different point-of-interests that users have visited. This app can also serve as a memory lane for users to remember their vacation.",
-  "main": "index.js",
-  "directories": {
-    "doc": "docs"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/DavidTangJK/David_TripPin.git"
-  },
-  "author": "davidtang",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/DavidTangJK/David_TripPin/issues"
-  },
-  "homepage": "https://github.com/DavidTangJK/David_TripPin#readme",
-  "dependencies": {
-    "@mapbox/search-js-web": "^1.0.0-beta.22"
-  }
-}
