+++ requirements.txt
@@ -1,0 +1,3 @@
+flask
+Flask-SQLAlchemy
+flask-login
