+++ website/auth.py
@@ -18,7 +18,7 @@ def login():
             if user.password == password:
                 flash('Logged in successfully!', category='success')
                 login_user(user, remember=True)
-                return redirect(url_for('views.index'))
+                return redirect(url_for('views.main_page'))
             else:
                 flash('Incorrect password, try again.', category='error')
         else:
@@ -62,6 +62,6 @@ def sign_up():
             db.session.commit()
             login_user(new_user, remember=True)
             flash('Account created!', category='success')
-            return redirect(url_for('views.index'))
+            return redirect(url_for('views.main_page'))
 
     return render_template("sign_up.html", user=current_user)
