+++ website/models.py
@@ -2,21 +2,19 @@ from . import db
 from flask_login import UserMixin
 from sqlalchemy.sql import func
 
-
-class Itinerary(db.Model):
-    __tablename__ = 'itinerary'
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(1000))
-    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
-    places = db.relationship('Places')
-
-class Places(db.Model):
+class Place(db.Model):
     __tablename__ = 'places'
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(1000), nullable=False)
     address = db.Column(db.String(5000), nullable=False)
     itinerary_id = db.Column(db.Integer, db.ForeignKey('itinerary.id'), nullable=False)
 
+class Itinerary(db.Model):
+    __tablename__ = 'itinerary'
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(1000))
+    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
+    places = db.relationship('Place')
 
 class User(db.Model, UserMixin):
     __tablename__ = 'users'
