+++ website/static/styles.css
@@ -68,10 +68,10 @@ body {
 
 #itinerary-panel {
   position: absolute;
-  top: 100px;
+  top: 150px;
   left: 10px;
-  width: 30%;
-  height: 50%;
+  width: 15%;
+  height: 20%;
   background: white;
   padding: 20px;
   overflow-y: auto;
